URI for data access request corrected, getModels and getOrganizations moved to DataCoreHandler
