+++ src/main/kotlin/org/ozwillo/dcimporter/web/Routes.kt
@@ -29,12 +29,10 @@ class Routes(
 
                 "/api").nest {
 
-            "/data_access_request".nest {
-                POST("/{request}", dataAccessRequestHandler::create)    //request = save for SAVED state, send for SENT state (meaning sent to validation task)
-                PUT("/{action}/{id}", dataAccessRequestHandler::update)     //action = valid (push "Valider" button, update state to VALIDATED) or reject (push "Annuler" button, update state to REFUSED) or send (in case of pre-filled form sent to validation) or save (in case or updating pre-filled form)
-                GET("/organizations", dataAccessRequestHandler::getAllOrganization)
-                GET("/model", dataAccessRequestHandler::getModels)
-                GET("/state/{state}", dataAccessRequestHandler::get)  //state = saved (list of saved pre-filled form) or sent (list of elements waiting for validation)
+            "/data-access".nest {
+                POST("", dataAccessRequestHandler::create)
+                PUT("/{id}/{action}", dataAccessRequestHandler::update)     //action = valid (push "Valider" button, update state to VALIDATED) or reject (push "Annuler" button, update state to REFUSED) or send (in case of pre-filled form sent to validation) or save (in case or updating pre-filled form)
+                GET("", dataAccessRequestHandler::get)
                 GET("/{id}", dataAccessRequestHandler::dataAccess)
             }
 
@@ -75,6 +73,8 @@ class Routes(
         "/dc".nest {
             POST("/type/{type}", datacoreHandler::createResourceWithOrganization)
             PUT("/type/{type}", datacoreHandler::updateResourceWithOrganization)
+            GET("/organizations", datacoreHandler::getAllOrganization)
+            GET("/models", datacoreHandler::getModels)
         }
 
         "/configuration".nest {
